memorystore: Return root-relative URLs.
Fixes an issue that happens when we store playlists hierarchically.
That is, given a playlist at <ManifestID>/<Rendition>.m3u8, the
segment at <ManifestID>/Rendition>/<Seq>.ts would have resulted
in the HLS player doing a lookup for

<ManifestID>/<ManifestID>/<Rendition>/<Seq>.ts

due to the relative path given by the playlist. Fix this by
making the path unconditionally root-relative; this results in
the correct prefix being applied to the request. Local URLs become

/source/<ManifestID>/<Rendition>/<Seq>.ts