PM: added sessionID field to winningTickets table
We want to be able to query all winning tickets received in the same
PM session. In the current Streamflow iteration we will attempt to
redeem winning tickets only once the video stream is over, which
practically means we will keep the same PM session for the entire video
stream, and it's likely O will receive multiple winning tickets in this
model. Therefore, we need to allow O to redeem all winning tickets from
a specific session (we can't just redeem all the tickets in the table
because O might have other streams in progress).

While in practice `sessionID` is `recipientRand`, we are choosing to
store it as a separate field to decouple O's Livepeer-specific code
from the internals of PM. This allows for a more generic and reusable
PM implementation, and maintaining a more stable O should any internal
PM changes happen.