PM: SenderNonce change to uint32
During testing of max values in storing winning tickets in the DB we
learned that we get a SQL library error trying to store uint64 max
values, so we chose the easy workaround of changing SenderNonce's range.

With unsigned 32 bits we still have 4.3B nonces, which if you assume
the case of one PM ticket per second still leaves us with a 136 years
of video we can stream on a single `recipientRand` value, so I believe
it should hold for now!