PM: WIP redeem winning tickets upon stream end
This is a first stab at introducing `Recipient` into the post-stream
flow.

We still need to make some changes to the `Recipient` API like changing
the redeem function to accept a list of sessionIDs, and then to mirror
that change in the `TicketStore` interface.

We're also waiting to sort out some of the other dependencies, e.g.
having ETH client implement the `Broker` interface before we can
complete the integration into production code.