tmanager: Distinguish between fatal and non-fatal errors.
Fatal errors likely indicate something wrong with the transcoder
itself and will lead to a disconnect. Non-fatal errors are more
likely to indicate an issue with the source stream or something
other than the transcoder, so maintain the transcoder within the
manager.

This also enables safer retries of failed transcodes; only fatal
errors are retries.