implement working windows gpu build
The recent change from `nvcc` to `clang` for the nvidia-flavored parts
of our ffmpeg build chain [1] basically gave us this feature for free.
All that was necessary was to add the mingw64 versions of the clang
compiler and it started happily producing binaries that interface with
CUDA on Windows. The rest was just refactoring:

I removed the Windows Docker build entirely. It was terrible, frequently
taking close to 90 minutes per build and frequently crashed. This change
implies that `docker/Dockerfile.build` and `docker/Dockerfile.build-linux`
can be combined back into one Dockerfile.build process; I'll open a tech
debt ticket for that but I don't think it's major enough to block this
merge.

Instead, we have a `.\windows-build.ps1` PowerShell script that takes
care of downloading MSYS2 using [Chocolatey](https://chocolatey.org/),
installing the necessary MSYS2/mingw64 packages, and running through
the build. It unpacks everything into a local .gitignored directory, so
it should be able to handle producing workspace-local binaries without
mucking with any system-level packages installed on the host.

I've tested this locally at my desk with a Windows T, Linux O, and macOS
B just to prove that it's possible. I have not done any kind of extensive
benchmarking work, but I did confirm that the `ja/lb` changes functioned
as expected with the appropriate Nvidia drivers.

This is the conclusion of a 10-month side project and I'm pretty stoked
😃

[1]: https://github.com/livepeer/go-livepeer/pull/1208