Accept Ethereum passphrase via file on disk rather than command line
When bringing up nodes on an actual network, `-ethPassword` can be
specified on the command line to supply the Ethereum account's
passphrase and allow unattended startup and operations. This works
for now, but it is very insecure. A attacker can compromise a system,
run `ps aux` and then see all command lines, including the go-livepeer
command line containing the Ethereum passphrase in clear text.

This adds the ability for the user to supply a filepath rather than a
cleartext password to the `-ethPassword` parameter. If the supplied
string is a valid file path, the passphrase will be read from the file,
otherwise the supplied string will be used as-is.

closes #1256