server: Cap transcode attempts.
Avoids thundering herd problems around repeated retries.

This will lower the success rate a bit, but is a more
realistic interpretation of satisfactory transcoding criteria.

The impact on the success rate is usually a few percent, < 5%.
When we see this type of impact, we're usually already below
99% anyway; for example going from ~95% -> ~92%. That
extra three percent would be for the segments that are retried
sometimes dozens of times. When that happens, it's also likely
that other segments are prone to be dropped, since a this point
it's a race between the O refresh and running out of Os.