discovery: Add versioning to filter out incompatible orchestrators.
This is network protocol version, rather than a node version or
smart contract version. Its primary use is for the broadcaster
to filter out orchestrators that may not support certain features.

Broadcasters will only work with orchestrators such that

    orchestrator.version >= broadcaster.version

Newer orchestrators should remain backwards compatible with older
broadcasters, but we won't assume forward compatibility vis-a-vis
older orchestrators working with newer broadcasters.

This is independent of the node or smart contract versions, both
of which could be upgraded without breaking changes that would
require a bump in network protocol version compatibility.