core,server: Run detection at the start of a LB session
Nvidia transcoder only allocates detection-related filters on the first
segment of a session.

If the B switches to a different O mid-stream on a segment that wasn't
scheduled for detection, we'll end up with a transcode session that
can't handle detection for subsequent segs.

So here we force-enable detection for the first segment of a new LB session.