drivers: Refactor package as a more generic and performant library (#2223)
* drivers: Make Size an optional property of files

The Filebase S3-compatible API does not return a
file size in the response for files greater than 255 bytes.

That caused a panic on s3.go:267 when reading recordings.

With this change, the problem won't happen anymore cause
we don't even try to dereference the ContentLength, but
hold it as a nilable pointer that should be checked wherever
it is used (right now: nowhere).

* drivers: Receive Reader instead of buffer on save

We want to use this lib for saving large files to ObjectStore
as well for the VOD tasks. Since they can be really large, we
don't want to have to buffer the whole files in memory for
saving. So changing this API to receive a Reader instead makes
more sense.

* core+server: Fix all usage of the drivers pkg

* core+server+verification: Fix tests

* drivers: Use S3 uploader for saving data

Avoid the need to buffer the whole input in memory before saving.
Will also be more performant for large files.

* drivers: Tune s3 uploader for better performance

Increase the concurrency and the part size so the uploads
are more performant in our production environment.

These are also more compatible with Storj's configuration
which saves files in 64MB chunks on the decentralized network,
so that part size also works well for them. But it still improves
upload performance to gcloud in around 10x as well.

* core: Use undeprecated session.NewSession

It returns an error so had to change a few
signatures.

* drivers: Create consts for uploader configs

Also add comments with context about how those numbers came up.

* Update CHANGELOG_PENDING