livepeer_cli: use O's currently registered ServiceURI as default address
When changing config options for the Orchestrator via livepeer_cli, the
public host:port of the node is requested for the configuration to be
applied successfully. This host:port address defaults to using the
ip-address of the node instead of using the global domain name that is
registered when setting up an O node. Use the registered Service URI as
the default option such that the entry can be left blank when
configuring options via the livepeer-cli.

Update CHANGELOG_PENDING.md

Co-authored-by: Rafał Leszko <rafal@livepeer.org>