Remove race test (#2852)
These tests are flaky and currently are failing in 80-90% or times, which makes our CI unusable. Probably it means that we have some race conditions, but it exists for such a long time that we should just accept this fact. We may revisit the race condition issues / tests when they start to become an issue in prod.