feat(ai): enable AI orchestrator discovery (#3004)
* feat(ai): enable AI orchestrator discovery

This commit incorporates the AIServiceRegistry contract address,
superseding the conventional ServiceRegistry contract address. This
strategic alteration streamlines the discovery process of AI Orchestrators
within the AI Subnet, thereby bolstering network accessibility and interaction.
While this approach serves as a swift workaround to enable the feature without
extensive code modification, it's important to note that it may disrupt the
existing transcoding discovery mechanism. We have to fix this if we want to
merge the two networks in the future.

* docs(ai): improve discovery documentation

This commit ensures that people are aware that they have to interact
with the `AIServiceRegistry` using their main Orch wallet.

* fix: fix 'AIServiceRegistry' devnet and testnet issue

This commit ensure that the hardcoded `AIServiceRegistry` contract
doesn't break the go-livepeer binary on local devnets or testnets.