feat(pricing): expand gateway max price to set per capability/model id (#3116)
This commit allows Gateways to specify the maximum pricing they are willing to pay for a given capability and model combination. Gateways use a single MaxPrice set at launch with the maxPricePerUnit flag, which serves as the default. Additionally, they can specify a JSON config using the maxPricePerCapability flag to set prices per capability and model ID. The maxPrice per capability can also be adjusted via the setBroadcastConfig endpoint of the CLI webserver.