ai/live: Add buffered segment read/writes (#3305)
Replaces most of the old pipe logic in the RTMP segmenter
which didn't really do anything except some logging.

This change allows for segment writes to an io.Writer (from RTMP)
to proceed without being blocked by reads (from trickle publish).
This avoids some potential back-pressure with MediaMTX in case
publishes are slow in going out for whatever reason.

This will also allow us to

1. have concurrent segments in-flight, and

2. track how many segments are in-flight as a proxy for "is this
    orchestrator slow" so we can switch orchestrators if necessary,
    since we should not have more than 1 segment in-flight (sometimes
    2 for a very brief overlap). This logic will be added in a follow-up PR.

The reader interface here also has a Clone method (also non-blocking)
which has the potential to simplify some of the reader handling around
payments, but payments is not changed right now.

Most of the code here was borrowed from the trickle server but generalized
into a proper reader/writer interface. Eventually we'll back-port this code into
the trickle server but it's not needed there right now.