ai/live: Slow orchestrator detection
Detect 'slow' orchs by keeping track of in-flight segments.
Count the difference between segments produced and segments completed.
There should only be ~1 segment in-flight concurrently.

Sometimes the beginning of the current segment may briefly overlap
with the end of the previous segment, so accommodate that by
checking for the second-to-last segment.