Disable payment check on the O's side (#3330)
* Disable payment check on the O's side

We encounter "insufficient balance" error from time to time on prod. This needs investigation, but since we're not using public Os, let's temporarily not stop steams even if not enough payment

* Disable payment check on the O's side

We encounter "insufficient balance" error from time to time on prod. This needs investigation, but since we're not using public Os, let's temporarily not stop steams even if not enough payment