ai/live: Close a segment if it will be dropped.
This is being polite to notify any subscribers that might be
waiting for the segment that the segment will not be around.

Not required if the segment is written normally, and subscribers
will still work fine without; it would just take longer to happen
when the segment drops out of the window of active segments.