docs(ai): improve VSCode AI launch.json (#3363)
This pull request ensures the `aiServiceRegistry` flag is added to the
example JSON, which is required to connect to the AI network. It also
improves the AI VSCode `launch.json` example by using a separate data
directory for the gateway and removes the unused `transcodingOptions` flag,
resulting in a cleaner and more organized configuration.