ai/live: Use events as a heartbeat. (#3382)
Terminate the stream if there are no incoming events for a period
of time - currently 30 seconds, but that can be tuned if 30 seconds
turns out to be too aggressive.

This allows the gateway to more quickly handle scenarios where the
runner completely explodes without a tear-down or stops responding
entirely. Otherwise the gateway would have to wait for the
orchestrator to expire idle trickle channels after roughly 5 minutes
before starting its own tear-down process.

Eventually this should be used as a signal to swap orchestrators
instead of terminating the incoming stream.