ai/live: Prevent a race condition with control channels (#3503)
* ai/live: Prevent a race condition with control channels

It is possible for us to receive concurrent connections for the
same stream key. For most things it is not an issue and we can
let stale connections tear themselves down eventually, but the
control API poses a special problem since it is indexed by stream
key, not request ID.

This could result in a stale connection shutting down the control
channel for an active connection.

To mitigate this, add a request ID to each control struct and
check the ID wherever we use the control struct. If we try to
overwrite an existing control struct, then shut the old one down.

Stomping over old control structs is not ideal and potentially
problematic but OK for now.

Some alternatives to consider later include:

- Keeping a list of control structs under the same stream key.
  Any control API calls would have to go out to each of them.

- Updating the API to access active streams only, eg by including
  the request ID in the control call.

* safer way to stop control channel

---------

Co-authored-by: Rafał Leszko <rafal@livepeer.org>