ai/live: Skip interleaving in segmenter (#3521)
The interleaver in the RTP segmenter has several issues [1].

Try to avoid those issues by skipping interleaving entirely.

Normally this is a terrible idea because it means that audio
and video could arrive relatively far apart leading to sync
issues during playback. However:

1. We aren't playing back the ingest stream. We control the
   demuxer here, and it doesn't care about the relative timings
   between tracks.

2. We do remux the output through a proper interleaver.
   So the sync should eventually fix itself on the runner side.

This is a bit of a YOLO move.

If things don't work out, it is easy to revert. If this does turn
out to be a good move, then we'll come back in here and clean up
a bunch of this now-unused code.

---

[1] Some of the issues are described in https://github.com/livepeer/go-livepeer/pull/3516

Worse things can happen when the segmenter's track queue
fills up. When a queue is full, packets are flushed out and a
high-watermark is set so we aren't writing packets out of order.

However, if another track is delayed (usually due to the RTP reorderer
buffering late / lost packets) then those delayed packets could arrive
at the segmenter *under* the high-water mark and be dropped. Sometimes
we cannot recover from this sequence at all, depending on the arrival
order.

Rather than try to fix interleaving, let's see if *not* interleaving
helps.