Fix make box syntax (#3535)
* Fix make box syntax

For some reason the existing syntax wasn't working on my machine so I have refactored the conditional a bit. I also had a conflict with us using the DOCKER_HOST variable, this is a built in config variable used by docker which I needed to set for go-livepeer to talk to docker desktop.

* revert to old syntax