ai/worker: Handle LOADING state on containers healthcheck (#3569)
* ai/worker: Be more lenient to unknown container statuses

We never know when we'll need to create another one, so let
them pass and just log.

* ai/worker: Add simple handling of LOADING state

* ai/worker: Fix startup timeouts for new LOADING state

* ai/worker: Make ERROR state restart immediately

We were waiting for 2 failures, but the ERROR state is
terminal and should be reacted immediately.

* TODO: reminder for another improvement

* ai/worker: Wait for LOADING state on container creation

* ai/worker: Pick up to date state for IDLE/LOADING checks

* ai/worker: Fix tests

* ai/worker: Reduce flakiness in tests

* ai/worker: Improve O startup time

We actually don't want to wait for LOADING to finish
on the createContainer function as that will block the
O from finishing startup. We should just initialize the
container state instead and let it go IDLE through the
normal logic on watchContainer.