Clone session for Realtime Video AI Processing (#3612)
The root reason for this change is that the "insufficient balance" error still sometimes happens on the Orchestrator. Upon investigation, I found that the problem is that the sess.Balance was shared between different streams causing one stream not accounting correctly. The simple fix is to always clone Balances (and other session params) before using it further in the Realtime Video pipeline.

Note that this change touches only Realtime Video (no other parts of the go-livepeer system).