ai/live: Avoid unnecessary swaps.
This fixes a race condition where a RTMP input shutting down
would trigger a swap if the connection was not yet fully
un-registered.

Specifically, the sequence of events is something like:

1. MediaMTX signals an input is closed

2. Segmenter is closed

3. Trickle publisher cancels the orchestrator context

4. Cancelling the orchestrator context triggers a swap

5. Input is marked as offline

Fix this by ensuring that we mark the input offline first
*before* cancelling the context.

We create a parent context for orchestrators and cancel that
after the input is marked as offline. Hence, having the trickle
publisher cancel the context is no longer necessary.

This usually affects RTMP; WHIP does not seem affected due to
a different order of operations: on disconnect, WHIP first
marks the input as offline, then closes the segmenter.

We *might* be able to adjust RTMP to have the same behavior as WHIP
but we generally want this change anyway, because early context
cancellation makes tear-down even more non-deterministic which is
undesirable. By making context cancellation happen later for client
disconnects, we retain more control of state transitions.