* Don't call stopProcessing on context cancellations.
The context is cancelled; nothing further needs to be done.
This avoids sending unnecessary events over the wire.

In general, the rule of thumb should be: if a chunk of code "owns"
the reason for triggering a swap, then it needs to call stopProcessing
(and cancel the context through that). Otherwise, the swap is caused
from something outside that chunk of code, so there's no need to call
stopProcessing again.

* Replace kickOrch with stopProcessing.

stopProcessing sends out an error report in addition to kicking the
orchestrator. If we're kicking an orch and initiating a swap, it is
helpful to understand why.

* Clean up some logging and fix a minor bug with a missed return.