ai/live: Allow context cancellation in trickle subscriber.
If there is a catastrophic issue with the runner and its published
channels are not shut down correctly (events and inference output),
then it is possible for the trickle subscriber Read() call to hang
until the server sweeps idle connections.

Fix this by adding a configuration struct to the subscriber and
pass in a context through that, among other fields such as the URL.

This way to pass contexts is a bit unusual, but how we use them within
the subscriber is already unusual, so highlight this oddity with a
slightly different type of API and some comments.

Also unit tests for context cancellation and server idle sweeps.