ai/live: Send trickle errors down to processStream (#3671)
This mitigates more tricky timing conditions. For example,
stream EOS and the "swap disabled" message come close enough
together that Metabase timestamps don't have enough granularity to
distinguish them, so they get displayed out of order.

Return any trickle errors down to processStream and preferentially
use those if available.

This also ensures we see a single error reason from trickle
when swapping or kicking, rather than potentially several. The
first error is generally closest to the root cause of an issue.

If an orchestrator will be swapped, we report the error.
If we are kicking the connection instead of swapping, then the
kickInput function reports the error.

Also see:
* https://github.com/livepeer/go-livepeer/pull/3666
* https://github.com/livepeer/go-livepeer/pull/3663