ai/live: Implement a watchdog for output segments. (#3711)
We have a timeout within copySegment but this does not cover
segment.Read() which can hang if the worker does not write any
HTTP headers as part of its trickle output stream.

This has happened a few times when runners fail to properly load
the model, or if there are other issues within the runner, egg
in pulling input segments.

This watchdog, which is reset on each segment using the existing
`outSegmentTimeout` timeout, will cover processing of the entire
segment, not just the segment body. If outSegmentTimeout is not
set, then a default of 30 seconds is used.