test: Add glog.FlushDaemon to goleak ignore list.
Since the move to golang 1.25 sometimes we get a failure
where unit tests exit while glog is in the middle of flushing via
glog.FlushDaemon. This puts the function in the middle of the call
stack, usually with some other OS functions on top.

Fix this by matching this specific function anywhere in the call stack,
(goleak.IgnoreAnyFunction) rather than just on top of the call stack
(goleak.IgnoreTopFunction).

We may encounter more of these in which case it might be justified to
stick everything under IgnoreAnyFunction but for now let's be cautious.