test: Fix more discovery race conditions (#3798)
This should hopefully clear up some of the sporadic errors we get
from CI around discovery since we switched to golang 1.25. We had
(and still have) a number of tests that unintentionally share state,
usually via modifying globals, and that causes them to stomp over
one another sometimes. In other places we had blocked goroutines
that weren't being terminated and running tests multiple times
(`-count 2`) triggered the goleak detector in other tests - those
were fun to track down.

* test: Add glog.FlushDaemon to goleak ignore list.

Since the move to golang 1.25 sometimes we get a failure
where unit tests exit while glog is in the middle of flushing via
glog.FlushDaemon. This puts the function in the middle of the call
stack, usually with some other OS functions on top.

Fix this by matching this specific function anywhere in the call stack,
(goleak.IgnoreAnyFunction) rather than just on top of the call stack
(goleak.IgnoreTopFunction).

We may encounter more of these in which case it might be justified to
stick everything under IgnoreAnyFunction but for now let's be cautious.

* discovery: Prevent blocked discovery requests.

* discovery/test: Prevent race condition with custom getOrchInfo

* discovery/test: Close hanging DB handle.

* discovery/test: Close DB orch pool context to tear down polling loop.

* server: Export NewBroadcastConfig() for tests.

Some unit tests were stomping over one another by using this global.

The struct does not allow us to set individual fields (and even so the
internals are a bit complex) so just do the lazy thing and export a
constructor function for tests to use.