trickle: Fix local subscriptions that start at a non-zero seq.
Also add a SetSeq helper.

We support subscribing to the next segment (seq -1) or the current
segment (seq -2). The default is -1.

However, if the local subscriber latches onto the server late, then the
current segment will have a nonzero sequence number. The local
calculation of the next segment did not take this into account. Fix this
by always settting the next seq based on the segment's own seq rather
than the local seq.