ai/live: Treat whep mpegts init errors as 404 (#3806)
This usually happens if the O swaps before producing any output and
should be benign. In this case the player should re-try the playback
and should eventually latch onto the next O.

However, the most appropriate error code to use here is tricky, since
technically this is a 500-class internal server error. However, 500s
produce alerts in our monitoring system, and alerting on benign errors
is poor on-call hygiene.

Use a 404 instead, simply because this is what existing clients using
MediaMTX are conditioned to expect if attempting to connect to a
stream that is not yet ready