ai/live: Prevent panic on concurrent map writes. (#3819)
We return a `status` map in the LiveVideoStatus HTTP handler then
mutate it to add the gateway status. This causes a panic if we
happen to have simultaneous accesses to this HTTP handler for the
same stream.

At some point we might want to think about merging the gateway
and status maps since they seem to carry similar per-stream data.
This would both make the bookkeeping simpler and avoid the need
for this (potentially expensive) copy on each request.