ai/live: Track latest input timestamp
Do this by attaching a timestamp to each MPEG-TS track. Most of the
changes here are a result of changing the segmenter API to use a new
SegmenterTrack struct that can more easily embed and test additional
fields [1] such as the last input timestamp.

This method allows us to cleanly extract the last input TS while
constructing the MediaState object.

The last input timestamp is stored in the original MPEG-TS 90khz
timebase for now, but converted to seconds before logging and Kafka.

[1] webrtc.TrackRemote is nearly impossible to cleanly mock, so
    make the SegmenterTrack take a RTPTrack so we can drop the
    existing TrackRemote in there alongside any additional fields.