ai/live: Measure gateway-to-orchestrator latency (#3829)
Tracks the last input timestamp and the last output timestamp.
Subtracting the input from the output allows for a continuous
measurement of gateway-to-orchestrator latency.

* ai/live: Track latest input timestamp

Do this by attaching a timestamp to each MPEG-TS track. Most of the
changes here are a result of changing the segmenter API to use a new
SegmenterTrack struct that can more easily embed and test additional
fields [1] such as the last input timestamp.

This method allows us to cleanly extract the last input TS while
constructing the MediaState object.

The last input timestamp is stored in the original MPEG-TS 90khz
timebase for now, but converted to seconds before logging and Kafka.

[1] webrtc.TrackRemote is nearly impossible to cleanly mock, so
    make the SegmenterTrack take a RTPTrack so we can drop the
    existing TrackRemote in there alongside any additional fields.

* ai/live: Track latest output timestamp

Use this to measure gateway-to-orchestrator latency continuously.
Surface this in both logs and Kafka metrics.

Introduce a global `media.OutputStats` to track the last seen MPEG-TS
timestamp per track. Spin up a MPEG-TS reader from the trickle
subscriber and publish each track's timestamp to that. For now, the
OutputStats timebase remains at the original 90khz; for logging and
Kafka it is converted to seconds.

When the track stats are calculated, also pull in the OutputStats.
Since we also have the last input timestamp, we can now measure latency.

* Embed RTPTrack into segmenterTrack